Update stylesheet.css
Was wondering if we could make this icon smaller, it looks stretched for me. Dunno if anyone else is having this problem.
Dunno if it's fix.
If it isn't I'm willing to upload a svg, and do wtv this people are doing here: https://github.com/eonpatapon/gnome-shell-extension-caffeine
Cause it seems to work.


Thanks!